module/requirements.txt
azure-iot-device
tcconfig
docker